Implement PacketTracker class
