+++ src/software/jetson_nano/BUILD
@@ -23,8 +23,8 @@ cc_library(
         "//proto:tbots_cc_proto",
         "//software/jetson_nano/redis",
         "//software/jetson_nano/services:motor",
-        "//software/jetson_nano/services:network",
         "//software/jetson_nano/services:power",
+        "//software/jetson_nano/services/network",
         "//software/logger:network_logger",
         "//software/util/scoped_timespec_timer",
         "//software/world:team",
