+++ src/software/jetson_nano/services/network/BUILD
@@ -1,0 +1,25 @@
+package(default_visibility = ["//visibility:public"])
+
+cc_library(
+    name = "network",
+    srcs = ["network.cpp"],
+    hdrs = ["network.h"],
+    deps = [
+        "//proto:tbots_cc_proto",
+        "//shared:robot_constants",
+        "//software/logger",
+        "//software/networking:threaded_proto_udp_listener",
+        "//software/networking:threaded_proto_udp_sender",
+        "@boost//:asio",
+        ":packet_tracker",
+    ],
+)
+
+cc_library(
+    name = "packet_tracker",
+    srcs = ["packet_tracker.cpp"],
+    hdrs = ["packet_tracker.h"],
+    deps = [
+        "//software/logger",
+    ],
+)
