+++ src/software/jetson_nano/services/network/network.h
@@ -7,11 +7,12 @@
 #include "proto/tbots_software_msgs.pb.h"
 #include "proto/world.pb.h"
 #include "shared/robot_constants.h"
+#include "software/jetson_nano/services/network/packet_tracker.h"
 #include "software/networking/threaded_proto_udp_listener.hpp"
 #include "software/networking/threaded_proto_udp_sender.hpp"
 
-constexpr float PACKET_LOSS_WARNING_THRESHOLD = 0.1f;
-constexpr uint8_t RECENT_PACKET_LOSS_PERIOD   = 100;
+// constexpr float PACKET_LOSS_WARNING_THRESHOLD = 0.1f;
+// constexpr uint8_t RECENT_PACKET_LOSS_PERIOD   = 100;
 
 class NetworkService
 {
@@ -65,4 +66,7 @@ class NetworkService
     // Queue of the sequence numbers of received worlds in the past
     // RECENT_PACKET_LOSS_PERIOD worlds
     std::queue<uint64_t> recent_world_seq_nums;
+
+    PacketTracker primitive_tracker = PacketTracker("primitive set");
+    PacketTracker world_tracker = PacketTracker("world");
 };
