+++ src/software/jetson_nano/thunderloop.h
@@ -9,7 +9,7 @@
 #include "software/jetson_nano/primitive_executor.h"
 #include "software/jetson_nano/redis/redis_client.h"
 #include "software/jetson_nano/services/motor.h"
-#include "software/jetson_nano/services/network.h"
+#include "software/jetson_nano/services/network/network.h"
 #include "software/jetson_nano/services/power.h"
 #include "software/logger/logger.h"
 #include "software/world/robot_state.h"
